@@ -17,12 +17,14 @@
   "dependencies": {
     "decentraland-ui": "^6.22.0",
     "ethers": "^6.13.5",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-redux": "^7.2.4",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
+    "react-redux": "^8.1.3",
+    "react-router-dom": "^7.9.3",
     "redux": "^5.0.1",
     "redux-logger": "^3.0.6",
-    "redux-saga": "^1.3.0"
+    "redux-saga": "^1.3.0",
+    "reselect": "^5.1.1"
   },
   "devDependencies": {
     "@babel/core": "^7.28.4",
@@ -31,13 +33,14 @@
     "@eslint/js": "^9.17.0",
     "@jridgewell/gen-mapping": "^0.3.13",
     "@testing-library/dom": "9.3",
-    "@testing-library/jest-dom": "5.16",
-    "@testing-library/react": "12.1",
-    "@testing-library/user-event": "14.4",
+    "@testing-library/jest-dom": "^6.8.0",
+    "@testing-library/react": "^14.3.1",
+    "@testing-library/user-event": "^14.6.1",
     "@types/jest": "^29.5.14",
-    "@types/react": "^17.0.83",
-    "@types/react-dom": "^17.0.26",
+    "@types/react": "^18.3.24",
+    "@types/react-dom": "^18.3.7",
     "@types/redux-logger": "^3.0.13",
+    "@types/reselect": "^2.0.27",
     "@vitejs/plugin-react": "^5.0.4",
     "@vitejs/plugin-react-swc": "^3.5.0",
     "eslint": "^9.36.0",
