@@ -1,14 +1,18 @@
-import ReactDOM from 'react-dom'
+import React from 'react'
+import { createRoot } from 'react-dom/client'
 import { Provider } from 'react-redux'
-
-import { App } from './components/App'
-import { store } from './modules/store'
+import { BrowserRouter } from 'react-router-dom'
 
 import 'decentraland-ui/lib/styles.css'
+import AppContainer from './components/App/App.container'
+import { store } from './modules/store'
 
-ReactDOM.render(
-  <Provider store={store}>
-    <App />
-  </Provider>,
-  document.getElementById('root'),
+createRoot(document.getElementById('root')!).render(
+  <React.StrictMode>
+    <Provider store={store}>
+      <BrowserRouter>
+        <AppContainer />
+      </BrowserRouter>
+    </Provider>
+  </React.StrictMode>,
 )
