@@ -1,29 +1,34 @@
 import { connect } from 'react-redux'
-import { connectWalletRequest, transferTokenRequest } from '../../modules/wallet/actions'
+import { App } from './App'
 import {
-  getAddress,
-  getBalance,
-  getError,
-  isConnected,
-  isConnecting,
-  isTransferring,
-} from '../../modules/wallet/selectors'
-import { RootState } from '../../modules/types'
+  selectAddress,
+  selectError,
+  selectIsConnected,
+  selectIsConnecting,
+  selectIsTransferring,
+  selectPrettyBalance,
+} from '../../modules/selectors/wallet'
+import {
+  connectWalletRequest,
+  transferTokenRequest,
+  refreshBalanceRequest,
+} from '../../modules/wallet/actions'
+import { RootState } from '../../modules/store'
 import { MapDispatch, MapDispatchProps, MapStateProps } from './App.types'
-import App from './App'
 
 const mapState = (state: RootState): MapStateProps => ({
-  address: getAddress(state),
-  balance: getBalance(state),
-  isConnected: isConnected(state),
-  isConnecting: isConnecting(state),
-  isTransferring: isTransferring(state),
-  error: getError(state),
+  address: selectAddress(state) ?? '',
+  balance: selectPrettyBalance(state),
+  isConnected: selectIsConnected(state),
+  isConnecting: selectIsConnecting(state),
+  isTransferring: selectIsTransferring(state),
+  error: selectError(state) ?? undefined,
 })
 
 const mapDispatch = (dispatch: MapDispatch): MapDispatchProps => ({
   onConnect: () => dispatch(connectWalletRequest()),
-  onTransfer: (to, amount) => dispatch(transferTokenRequest({ to, amount })),
+  onTransfer: (to: string, amount: string) => dispatch(transferTokenRequest({ to, amount })),
+  onRefreshBalance: () => dispatch(refreshBalanceRequest()),
 })
 
 export default connect(mapState, mapDispatch)(App)
