@@ -2,6 +2,7 @@ import { Action, Dispatch } from 'redux'
 import {
   ConnectWalletRequestAction,
   TransferTokenRequestAction,
+  RefreshBalanceRequestAction,
 } from '../../modules/wallet/actions'
 
 export type Props = {
@@ -14,11 +15,14 @@ export type Props = {
   error: string | null
   onConnect: () => void
   onTransfer: (to: string, amount: string) => void
+  onRefreshBalance: () => void
 }
 
 export type MapStateProps = Pick<
   Props,
   'address' | 'balance' | 'isConnected' | 'isConnecting' | 'isTransferring' | 'error'
 >
-export type MapDispatchProps = Pick<Props, 'onConnect' | 'onTransfer'>
-export type MapDispatch = Dispatch<ConnectWalletRequestAction | TransferTokenRequestAction | Action>
+export type MapDispatchProps = Pick<Props, 'onConnect' | 'onTransfer' | 'onRefreshBalance'>
+export type MapDispatch = Dispatch<
+  ConnectWalletRequestAction | TransferTokenRequestAction | RefreshBalanceRequestAction | Action
+>
