@@ -3,6 +3,7 @@
 import React, { PropsWithChildren } from 'react'
 import { render, RenderOptions } from '@testing-library/react'
 import { Provider } from 'react-redux'
+import { BrowserRouter } from 'react-router-dom'
 import { applyMiddleware, combineReducers } from 'redux'
 import createSagaMiddleware from 'redux-saga'
 import { walletReducer } from '../modules/wallet/reducer'
@@ -30,7 +31,11 @@ export function renderWithProviders(
   }: { preloadedState?: Partial<TestRootState>; store?: any } & Omit<RenderOptions, 'wrapper'> = {},
 ) {
   function Wrapper({ children }: PropsWithChildren<{}>) {
-    return <Provider store={store}>{children}</Provider>
+    return (
+      <Provider store={store}>
+        <BrowserRouter>{children}</BrowserRouter>
+      </Provider>
+    )
   }
   return { store, ...render(ui, { wrapper: Wrapper, ...renderOptions }) }
 }
